Document + create AddrResolver entity + move rsk registrar e2e + use inline abis
