+++ package.json
@@ -31,6 +31,7 @@
   "homepage": "https://github.com/rsksmart/rns-sdk#readme",
   "devDependencies": {
     "@rsksmart/erc677": "^1.0.2",
+    "@rsksmart/rns-registry": "^1.0.4",
     "@rsksmart/rns-resolver": "^2.0.0",
     "@rsksmart/rns-resolver.js": "^1.0.1",
     "@types/jest": "^26.0.23",
@@ -52,7 +53,6 @@
   },
   "dependencies": {
     "@ensdomains/eth-ens-namehash": "^2.0.15",
-    "@rsksmart/rns-registry": "^1.0.4",
     "ethers": "^5.5.1",
     "js-sha3": "^0.8.0"
   }
