+++ src/index.ts
@@ -1,0 +1,4 @@
+export { RNS } from './RNS'
+export { AddrResolver } from './AddrResolver'
+export { RSKRegistrar } from './RSKRegistrar'
+export * from './hash'
