+++ test/rns.test.ts
@@ -17,30 +17,6 @@ describe('RNS SDK', () => {
     expect(await rnsRegistryContract.owner(hashDomain(TEST_TARINGA_SUBDOMAIN))).toEqual(TEST_ADDRESS)
   })
 
-  test('set addr for taringa.rsk', async () => {
-    const { taringaOwner, rnsRegistryContract, addrResolverContract, registerSubdomain } = await deployRNS()
-    await registerSubdomain()
-
-    const rns = new RNS(rnsRegistryContract.address, taringaOwner)
-
-    const tx = await rns.setAddr(TEST_TARINGA_SUBDOMAIN, TEST_ADDRESS)
-    await tx.wait()
-
-    expect(await addrResolverContract.addr(hashDomain(TEST_TARINGA_SUBDOMAIN))).toEqual(TEST_ADDRESS)
-  })
-
-  test('addr for taringa.rsk', async () => {
-    const { taringaOwner, rnsRegistryContract, addrResolverContract, registerSubdomain } = await deployRNS()
-    await registerSubdomain()
-    await sendAndWait(addrResolverContract.setAddr(hashDomain(TEST_TARINGA_SUBDOMAIN), TEST_ADDRESS))
-
-    const rns = new RNS(rnsRegistryContract.address, taringaOwner)
-
-    const addressResolved = await rns.addr(TEST_TARINGA_SUBDOMAIN)
-
-    expect(addressResolved).toEqual(TEST_ADDRESS)
-  })
-
   test('set taringa.rsk owner', async () => {
     const { taringaOwner, rnsRegistryContract } = await deployRNS()
